Sat, Dec 21, 2024, 8:02 AM +01:00

+++ week6/pset/sentimental-credit/credit.py
@@ -70,7 +70,7 @@ def square(base, exponent):
 
 
 def starting(number, len, digit):
-    return (number - number % square(10, digit - len)) / square(10, digit - len)
+    return int((number - number % square(10, digit - len)) / square(10, digit - len))
 
 
 main()
